requirements.txt
alembic==1.12.1
beautifulsoup4==4.12.2
black==23.11.0
blinker==1.7.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
envparse==0.2.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
